+++ module/data/actor/npc.js
@@ -5,6 +5,7 @@ import {
   calculateHealthMax,
   calculateMeleeDamageBonusFormula,
   clampSanityRitualValue,
+  cleanDerivedNumber,
   computeEquippedArmorProtection,
   initializeSanityIfUnset,
   prepareStatisticsX5,
@@ -24,8 +25,9 @@ export default class NPCData extends CharacterData {
       ...HumanSkillsActorData.defineSchema(),
       schemaVersion: new NumberField({ initial: 2.0 }),
       sanity: new SchemaField({
-        value: new NumberField({ initial: 100 }),
-        currentBreakingPoint: new NumberField({ initial: 101 }),
+        value: new NumberField({ initial: 100, min: 0, integer: true }),
+        max: new NumberField({ initial: 99, min: 0, integer: true }),
+        currentBreakingPoint: new NumberField({ initial: 101, integer: true }),
       }),
       biography: new SchemaField({
         profession: new StringField({ initial: "" }),
@@ -38,13 +40,28 @@ export default class NPCData extends CharacterData {
 
   /** @inheritdoc */
   prepareDerivedData() {
-    prepareStatisticsX5(this.statistics);
+    const sourceStatistics = this.parent?._source?.system?.statistics;
+    prepareStatisticsX5(this.statistics, sourceStatistics);
     initializeSanityIfUnset(this.sanity, this.statistics);
-    this.sanity.max = 99 - this.skills.unnatural.proficiency;
-    this.wp.max = this.statistics.pow.value;
-    this.health.max = calculateHealthMax(this.statistics);
+    this.sanity.max = cleanDerivedNumber(
+      this,
+      "sanity.max",
+      99 - this.skills.unnatural.proficiency,
+    );
+    this.wp.max = cleanDerivedNumber(
+      this,
+      "wp.max",
+      this.statistics.pow.effectiveValue ?? this.statistics.pow.value,
+    );
+    this.health.max = cleanDerivedNumber(
+      this,
+      "health.max",
+      calculateHealthMax(this.statistics, sourceStatistics),
+    );
     this.statistics.str.meleeDamageBonusFormula =
-      calculateMeleeDamageBonusFormula(this.statistics.str.value);
+      calculateMeleeDamageBonusFormula(
+        this.statistics.str.effectiveValue ?? this.statistics.str.value,
+      );
     removeLegacyRitualSkill(this);
     clampSanityRitualValue(this.sanity);
     this.health.protection = computeEquippedArmorProtection(this.parent.items);