+++ module/utils/actor-submit.js
@@ -4,13 +4,12 @@
  * @returns {void}
  */
 export function flagUnnaturalResourceMaxUpdates(submittedData, actor) {
+  const sourceStats = actor._source?.system?.statistics;
+
   submittedData.system.wp.maxNeedsUpdate =
-    submittedData.system?.statistics.pow.value !==
-    actor.system.statistics.pow.value;
+    submittedData.system?.statistics?.pow?.value !== sourceStats?.pow?.value;
 
   submittedData.system.health.maxNeedsUpdate =
-    submittedData.system?.statistics.str.value !==
-      actor.system.statistics.str.value ||
-    submittedData.system?.statistics.con.value !==
-      actor.system.statistics.con.value;
+    submittedData.system?.statistics?.str?.value !== sourceStats?.str?.value ||
+    submittedData.system?.statistics?.con?.value !== sourceStats?.con?.value;
 }